refactor: Score 클래스 삭제
