+++ src/main/java/blackjack/domain/player/Participant.java
@@ -1,7 +1,5 @@
 package blackjack.domain.player;
 
-import blackjack.domain.result.Score;
-
 public class Participant extends Player {
 
     public Participant(final String name) {
@@ -17,6 +15,6 @@ public class Participant extends Player {
 
     @Override
     public boolean canAddCard() {
-        return new Score(this).CanAddPlayerCard();
+        return !isBust();
     }
 }
