+++ src/test/java/blackjack/domain/player/PlayerTest.java
@@ -1,7 +1,13 @@
 package blackjack.domain.player;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Denomination;
+import blackjack.domain.card.Suit;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -33,5 +39,54 @@ public class PlayerTest {
                 })
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("[ERROR] 이름에는 특수문자가 들어갈 수 없습니다.");
+    }
+
+    @Test
+    @DisplayName("플레이어가 블랙잭이 아닐 때 확인")
+    void isNotBlackJackScoreTest() {
+        Player player = new Player("배카라") {
+            @Override
+            public boolean canAddCard() {
+                return false;
+            }
+        };
+
+        player.addCard(new Card(Denomination.ACE, Suit.HEART));
+        player.addCard(new Card(Denomination.ACE, Suit.DIAMOND));
+
+        assertThat(player.isBlackJack()).isFalse();
+    }
+
+    @Test
+    @DisplayName("플레이어가 블랙잭인지 확인")
+    void isBlackJackTest() {
+        Player player = new Player("배카라") {
+            @Override
+            public boolean canAddCard() {
+                return false;
+            }
+        };
+
+        player.addCard(new Card(Denomination.ACE, Suit.HEART));
+        player.addCard(new Card(Denomination.TEN, Suit.DIAMOND));
+
+        assertThat(player.isBlackJack()).isTrue();
+    }
+
+    @Test
+    @DisplayName("플레이어가 블랙잭 카드 사이즈가 아닐 때 확인")
+    void isNotBlackJackSizeTest() {
+        Player player = new Player("배카라") {
+            @Override
+            public boolean canAddCard() {
+                return false;
+            }
+        };
+
+        player.addCard(new Card(Denomination.TEN, Suit.HEART));
+        player.addCard(new Card(Denomination.THREE, Suit.HEART));
+        player.addCard(new Card(Denomination.EIGHT, Suit.HEART));
+
+        assertThat(player.isBlackJack()).isFalse();
     }
 }
